Move lastN from Array to StreamD
Re-export it back in Array
